+++ lib/validators/email.ex
@@ -10,8 +10,8 @@ defmodule EctoCommons.EmailValidator do
     their `type="email"` input fields. This is the default as it corresponds to most use-cases. It is quite strict
     without being too narrow. It does not support unicode emails though. If you need better internationalization,
     please use the `:pow` check as it is more flexible with international emails. Defaults to enabled.
-  - `:burner`: Checks if the email given is a burner email provider. When enabled, will reject temporary
-    email providers. Defaults to disabled.
+  - `:burner`: Checks if the email given is a burner email provider (uses the `Burnex` lib under the hood).
+    When enabled, will reject temporary email providers. Defaults to disabled.
   - `:pow`: Checks the email using the [`pow`](https://hex.pm/packages/pow) logic. Defaults to disabled.
     The rules are the following:
     - Split into local-part and domain at last `@` occurrence
@@ -27,9 +27,18 @@ defmodule EctoCommons.EmailValidator do
     - Domain should;
       - be at most 255 octets
       - only have letters, digits, hyphen, and dots
-      - do not start or end with hyphen or dot
-      - can be an IPv4 or IPv6 address
+
     Unicode characters are permitted in both local-part and domain.
+
+    The implementation is based on [RFC 3696](https://tools.ietf.org/html/rfc3696#section-3).
+    IP addresses are not allowed as per the RFC 3696 specification: "The domain name can also be
+    replaced by an IP address in square brackets, but that form is strongly discouraged except
+    for testing and troubleshooting purposes.".
+
+    You're invited to compare the tests to see the difference between the `:html_input`
+    check and the `:pow` check. `:pow` is better suited for i18n and is more correct
+    in regards to the email specification but will allow valid emails many systems don't
+    manage correctly. `:html_input` is more basic but should be OK for most common use-cases.
 
   ## Example:
 
@@ -67,13 +76,9 @@ defmodule EctoCommons.EmailValidator do
   # credo:disable-for-next-line Credo.Check.Readability.MaxLineLength
   @email_regex ~r/^[a-zA-Z0-9.!#$%&'*+\/=?^_`{|}~-]+@[a-zA-Z0-9](?:[a-zA-Z0-9-]{0,61}[a-zA-Z0-9])?(?:\.[a-zA-Z0-9](?:[a-zA-Z0-9-]{0,61}[a-zA-Z0-9])?)*$/
 
-  # credo:disable-for-next-line Credo.Check.Readability.MaxLineLength
-  @ipv6_regex ~r/(([0-9a-fA-F]{1,4}:){7,7}[0-9a-fA-F]{1,4}|([0-9a-fA-F]{1,4}:){1,7}:|([0-9a-fA-F]{1,4}:){1,6}:[0-9a-fA-F]{1,4}|([0-9a-fA-F]{1,4}:){1,5}(:[0-9a-fA-F]{1,4}){1,2}|([0-9a-fA-F]{1,4}:){1,4}(:[0-9a-fA-F]{1,4}){1,3}|([0-9a-fA-F]{1,4}:){1,3}(:[0-9a-fA-F]{1,4}){1,4}|([0-9a-fA-F]{1,4}:){1,2}(:[0-9a-fA-F]{1,4}){1,5}|[0-9a-fA-F]{1,4}:((:[0-9a-fA-F]{1,4}){1,6})|:((:[0-9a-fA-F]{1,4}){1,7}|:)|fe80:(:[0-9a-fA-F]{0,4}){0,4}%[0-9a-zA-Z]{1,}|::(ffff(:0{1,4}){0,1}:){0,1}((25[0-5]|(2[0-4]|1{0,1}[0-9]){0,1}[0-9])\.){3,3}(25[0-5]|(2[0-4]|1{0,1}[0-9]){0,1}[0-9])|([0-9a-fA-F]{1,4}:){1,4}:((25[0-5]|(2[0-4]|1{0,1}[0-9]){0,1}[0-9])\.){3,3}(25[0-5]|(2[0-4]|1{0,1}[0-9]){0,1}[0-9]))/
-  @ipv4_regex ~r/((25[0-5]|(2[0-4]|1{0,1}[0-9]){0,1}[0-9])\.){3,3}(25[0-5]|(2[0-4]|1{0,1}[0-9]){0,1}[0-9])/
-
   def validate_email(%Ecto.Changeset{} = changeset, field, opts \\ []) do
     validate_change(changeset, field, {:email, opts}, fn _, value ->
-      checks = Keyword.get(opts, :checks, [:pow])
+      checks = Keyword.get(opts, :checks, [:html_input])
 
       # credo:disable-for-lines:6 Credo.Check.Refactor.Nesting
       Enum.reduce(checks, [], fn check, errors ->
@@ -89,11 +94,11 @@ defmodule EctoCommons.EmailValidator do
   @spec do_validate_email(String.t(), atom()) :: :ok | {:error, String.t()}
   defp do_validate_email(email, :burner) do
     case Burnex.is_burner?(email) do
-      true ->
-        {:error, "uses a forbidden provider"}
-
       false ->
         :ok
+
+      true ->
+        {:error, "uses a forbidden provider"}
     end
   end
 
@@ -111,15 +116,15 @@ defmodule EctoCommons.EmailValidator do
   end
 
   # The code below is copied and adapted from the [pow](https://hex.pm/packages/pow) package
-  # with a few fixes on the domain part.
+  # We just don't want to import the whole `pow` package as a dependency.
   defp pow_validate_email(email) do
-    [domain | rest] =
+    [domain | local_parts] =
       email
       |> String.split("@")
       |> Enum.reverse()
 
     local_part =
-      rest
+      local_parts
       |> Enum.reverse()
       |> Enum.join("@")
 
@@ -127,19 +132,20 @@ defmodule EctoCommons.EmailValidator do
       String.length(local_part) > 64 -> {:error, "local-part too long"}
       String.length(domain) > 255 -> {:error, "domain too long"}
       local_part == "" -> {:error, "invalid format"}
+      local_part_only_quoted?(local_part) -> validate_domain(domain)
       true -> pow_validate_email(local_part, domain)
     end
   end
 
   defp pow_validate_email(local_part, domain) do
-    sanitized_local_part = remove_quotes_from_local_part(local_part)
+    sanitized_local_part =
+      local_part
+      |> remove_comments()
+      |> remove_quotes_from_local_part()
 
     cond do
-      local_part_only_quoted?(local_part) ->
-        validate_domain(domain)
-
-      local_part_consecutive_dots?(sanitized_local_part) ->
-        {:error, "consecutive dots in local-part"}
+      local_part_consective_dots?(sanitized_local_part) ->
+        {:error, "consective dots in local-part"}
 
       local_part_valid_characters?(sanitized_local_part) ->
         validate_domain(domain)
@@ -149,28 +155,66 @@ defmodule EctoCommons.EmailValidator do
     end
   end
 
+  defp local_part_only_quoted?(local_part),
+    do: local_part =~ ~r/^"[^\"]+"$/
+
   defp remove_quotes_from_local_part(local_part),
     do: Regex.replace(~r/(^\".*\"$)|(^\".*\"\.)|(\.\".*\"$)?/, local_part, "")
 
-  defp local_part_only_quoted?(local_part), do: local_part =~ ~r/^"[^\"]+"$/
+  defp remove_comments(any),
+    do: Regex.replace(~r/(^\(.*\))|(\(.*\)$)?/, any, "")
 
-  defp local_part_consecutive_dots?(local_part), do: local_part =~ ~r/\.\./
+  defp local_part_consective_dots?(local_part),
+    do: local_part =~ ~r/\.\./
 
   defp local_part_valid_characters?(sanitized_local_part),
-    do: sanitized_local_part =~ ~r<^[\p{L}0-9!#$%&'*+-/=?^_`{|}~\.]+$>u
+    do: sanitized_local_part =~ ~r<^[\p{L}\p{M}0-9!#$%&'*+-/=?^_`{|}~\.]+$>u
 
   defp validate_domain(domain) do
-    cond do
-      String.first(domain) == "-" -> {:error, "domain begins with hyphen"}
-      String.first(domain) == "." -> {:error, "domain begins with a dot"}
-      String.last(domain) == "-" -> {:error, "domain ends with hyphen"}
-      String.last(domain) == "." -> {:error, "domain ends with a dot"}
-      domain =~ ~r/^[\p{L}0-9-\.]+$/u -> :ok
-      domain =~ @ipv6_regex -> :ok
-      domain =~ @ipv4_regex -> :ok
-      true -> {:error, "invalid domain"}
+    sanitized_domain = remove_comments(domain)
+
+    labels =
+      sanitized_domain
+      |> remove_comments()
+      |> String.split(".")
+
+    labels
+    |> validate_tld()
+    |> validate_dns_labels()
+  end
+
+  defp validate_tld(labels) do
+    labels
+    |> List.last()
+    |> Kernel.=~(~r/^[0-9]+$/)
+    |> case do
+      true -> {:error, "tld cannot be all-numeric"}
+      false -> {:ok, labels}
     end
   end
+
+  defp validate_dns_labels({:ok, labels}) do
+    Enum.reduce_while(labels, :ok, fn
+      label, :ok -> {:cont, validate_dns_label(label)}
+      _label, error -> {:halt, error}
+    end)
+  end
+
+  defp validate_dns_labels({:error, error}), do: {:error, error}
+
+  defp validate_dns_label(label) do
+    cond do
+      label == "" -> {:error, "dns label is too short"}
+      String.length(label) > 63 -> {:error, "dns label too long"}
+      String.first(label) == "-" -> {:error, "dns label begins with hyphen"}
+      String.last(label) == "-" -> {:error, "dns label ends with hyphen"}
+      dns_label_valid_characters?(label) -> :ok
+      true -> {:error, "invalid characters in dns label"}
+    end
+  end
+
+  defp dns_label_valid_characters?(label),
+    do: label =~ ~r/^[\p{L}\p{M}0-9-]+$/u
 
   defp message(opts, key \\ :message, default) do
     Keyword.get(opts, key, default)