+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Changelog
 
 ## Upcoming
+- Ecto Validators:
+  - `validate_email`: Fix default checks to `:html_input` instead of `:pow` and
+     adds tests to confirm this is always the case to match documentation
+  - `validate_email`: Update `:pow` validation with upstream changes
 - Misc:
   - moved from Travis CI to GitHub Actions